+++ apps/storybook/src/components/layout/baseLayout.tsx
@@ -6,7 +6,8 @@ import { useOrderlyConfig } from "../../hooks/useOrderlyConfig";
 type BaseLayoutProps = {
   children: React.ReactNode;
   initialMenu?: string;
-} & Pick<ScaffoldProps, "classNames">;
+  classNames?: ScaffoldProps["classNames"];
+};
 
 export const BaseLayout: FC<BaseLayoutProps> = (props) => {
   const { onRouteChange } = useNav();