+++ packages/i18n/src/index.ts
@@ -14,4 +14,4 @@ export { useLocaleContext, type Language } from "./context";
 export * from "./types";
 export * from "./constant";
 export * from "./locale/en";
-export { parseI18nLang } from "./utils";
+export * from "./utils";