+++ packages/ui-orders/src/utils/util.ts
@@ -220,3 +220,12 @@ export function calcBracketRoiAndPnL(order: API.AlgoOrderExt) {
     },
   };
 }
+
+export function areDatesEqual(date1?: Date, date2?: Date): boolean {
+  if (!date1 || !date2) return false;
+  return (
+    date1.getFullYear() === date2.getFullYear() &&
+    date1.getMonth() === date2.getMonth() &&
+    date1.getDate() === date2.getDate()
+  );
+}