+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @orderly.network/hooks
 
+## 2.1.1
+
+### Patch Changes
+
+- publish
+- 0b52044b6: publish
+- Updated dependencies
+- Updated dependencies [0b52044b6]
+  - @orderly.network/core@2.1.1
+  - @orderly.network/default-evm-adapter@2.1.1
+  - @orderly.network/default-solana-adapter@2.1.1
+  - @orderly.network/net@2.1.1
+  - @orderly.network/perp@4.1.1
+  - @orderly.network/types@2.1.1
+  - @orderly.network/utils@2.1.1
+  - @orderly.network/web3-provider-ethers@2.1.1
+
+## 2.1.1-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/core@2.1.1-alpha.0
+  - @orderly.network/default-evm-adapter@2.1.1-alpha.0
+  - @orderly.network/default-solana-adapter@2.1.1-alpha.0
+  - @orderly.network/net@2.1.1-alpha.0
+  - @orderly.network/perp@4.1.1-alpha.0
+  - @orderly.network/types@2.1.1-alpha.0
+  - @orderly.network/utils@2.1.1-alpha.0
+  - @orderly.network/web3-provider-ethers@2.1.1-alpha.0
+
 ## 2.1.0
 
 ### Minor Changes