+++ packages/walletconnect/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @orderly.network/web3-modal
 
+## 2.1.1
+
+### Patch Changes
+
+- publish
+- 0b52044b6: publish
+- Updated dependencies
+- Updated dependencies [0b52044b6]
+  - @orderly.network/hooks@2.1.1
+
+## 2.1.1-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/hooks@2.1.1-alpha.0
+
 ## 2.1.0
 
 ### Minor Changes